ErrorHandler: Move to MPDActivity.
