+++ MPDroid/src/main/java/com/namelessdev/mpdroid/NowPlayingActivity.java
@@ -41,8 +41,6 @@ public class NowPlayingActivity extends MPDActivity {
 
     private static final String TAG = "NowPlayingActivity";
 
-    private ErrorHandler mErrorHandler;
-
     private boolean mIsDualPaneMode;
 
     private ViewPager mNowPlayingPager;
@@ -186,15 +184,6 @@ public class NowPlayingActivity extends MPDActivity {
         return itemHandled;
     }
 
-    /**
-     * Dispatch onPause() to fragments.
-     */
-    @Override
-    protected void onPause() {
-        mErrorHandler.stop();
-        super.onPause();
-    }
-
     @Override
     public boolean onPrepareOptionsMenu(final Menu menu) {
         super.onPrepareOptionsMenu(menu);
@@ -228,21 +217,6 @@ public class NowPlayingActivity extends MPDActivity {
         menu.findItem(R.id.GMM_Consume).setChecked(mpdStatus.isConsume());
 
         return true;
-    }
-
-    /**
-     * Dispatch onResume() to fragments.  Note that for better inter-operation
-     * with older versions of the platform, at the point of this call the
-     * fragments attached to the activity are <em>not</em> resumed.  This means
-     * that in some cases the previous state may still be saved, not allowing
-     * fragment transactions that modify the state.  To correctly interact
-     * with fragments in their proper state, you should instead override
-     * {@link #onResumeFragments()}.
-     */
-    @Override
-    protected void onResume() {
-        super.onResume();
-        mErrorHandler = new ErrorHandler(this);
     }
 
     public void showQueue() {
