+++ MPDroid/src/main/java/com/namelessdev/mpdroid/library/LibraryTabsSettings.java
@@ -18,7 +18,6 @@ package com.namelessdev.mpdroid.library;
 
 import com.mobeta.android.dslv.DragSortController;
 import com.mobeta.android.dslv.DragSortListView;
-import com.namelessdev.mpdroid.ErrorHandler;
 import com.namelessdev.mpdroid.R;
 import com.namelessdev.mpdroid.adapters.SeparatedListAdapter;
 import com.namelessdev.mpdroid.adapters.SeparatedListDataBinder;
@@ -39,8 +38,6 @@ import java.util.List;
 public class LibraryTabsSettings extends PreferenceActivity {
 
     private SeparatedListAdapter mAdapter;
-
-    private ErrorHandler mErrorHandler;
 
     private ArrayList<Object> mTabList;
 
@@ -121,18 +118,6 @@ public class LibraryTabsSettings extends PreferenceActivity {
                 return super.onOptionsItemSelected(item);
         }
 
-    }
-
-    @Override
-    protected void onPause() {
-        mErrorHandler.stop();
-        super.onPause();
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        mErrorHandler = new ErrorHandler(this);
     }
 
     private void refreshTable() {
